+++ tests/Config/YamlConfigTest.php
@@ -3,11 +3,12 @@
 namespace Test\OctoLab\Cilex\Config;
 
 use OctoLab\Cilex\Config\Loader\YamlFileLoader;
+use OctoLab\Cilex\Config\Parser\SymfonyYamlParser;
 use OctoLab\Cilex\Config\YamlConfig;
 use Symfony\Component\Config\FileLocator;
 
 /**
- * phpunit src/Tests/Config/YamlConfigTest.php
+ * phpunit tests/Config/YamlConfigTest.php
  *
  * @author Kamil Samigullin <kamil@samigullin.info>
  */
@@ -19,7 +20,7 @@ class YamlConfigTest extends \PHPUnit_Framework_TestCase
      */
     public function throwDomainException()
     {
-        $config = new YamlConfig(new YamlFileLoader(new FileLocator()));
+        $config = new YamlConfig(new YamlFileLoader(new FileLocator(), new SymfonyYamlParser()));
         $config->load('not_yaml.file', true);
     }
 }